extend edit bookmark suite. Parameterize getBookmark() with authentication method, basic auth by default

+++ ownCloudTests/Login/EditBookmarkTests.swift
@@ -25,20 +25,51 @@ class EditBookmarkTests: XCTestCase {
 	}
 
 	/*
-	* PASSED if: URL and Delete Auth Data displayed
-	*/
-	func testCheckInitialEditViewAuth () {
+	 * PASSED if: URL and Delete Auth Data displayed
+	 */
+	func testCheckInitialViewEditBasicAuth () {
 
 		if let bookmark: OCBookmark = UtilsTests.getBookmark() {
 
 			OCBookmarkManager.shared.addBookmark(bookmark)
 			UtilsTests.refreshServerList()
 
+			//Actions
 			EarlGrey.select(elementWithMatcher: grey_accessibilityID("server-bookmark-cell")).perform(grey_swipeFastInDirection(.left))
 			EarlGrey.select(elementWithMatcher: grey_text("Edit".localized)).perform(grey_tap())
 
 			//Assert
 			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-url-url")).assert(grey_sufficientlyVisible())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-auth-data-delete")).assert(grey_sufficientlyVisible())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-password")).assert(grey_sufficientlyVisible())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-username")).assert(grey_sufficientlyVisible())
+
+			//Reset status
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("cancel")).perform(grey_tap())
+			UtilsTests.deleteAllBookmarks()
+			UtilsTests.refreshServerList()
+		}
+	}
+
+	/*
+	 * PASSED if: URL, Delete Auth Data and Authentication message displayed if OAuth2
+	 */
+	func testCheckInitialViewEditOAuth2 () {
+
+		if let bookmark: OCBookmark = UtilsTests.getBookmark(authenticationMethod: OCAuthenticationMethodIdentifier.oAuth2) {
+
+			OCBookmarkManager.shared.addBookmark(bookmark)
+			UtilsTests.refreshServerList()
+
+			//Actions
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("server-bookmark-cell")).perform(grey_swipeFastInDirection(.left))
+			EarlGrey.select(elementWithMatcher: grey_text("Edit".localized)).perform(grey_tap())
+
+			//Assert
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-url-url")).assert(grey_sufficientlyVisible())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-username")).assert(grey_notVisible())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-password")).assert(grey_notVisible())
+			EarlGrey.select(elementWithMatcher: grey_text("Authenticated via OAuth2".localized)).assert(grey_sufficientlyVisible())
 			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-auth-data-delete")).assert(grey_sufficientlyVisible())
 
 			//Reset status
@@ -49,8 +80,8 @@ class EditBookmarkTests: XCTestCase {
 	}
 
 	/*
-	* PASSED if: Server name has change to "New name"
-	*/
+	 * PASSED if: Server name has change to "New name"
+	 */
 	func testCheckEditServerName () {
 
 		let expectedServerName = "New name"
@@ -60,6 +91,7 @@ class EditBookmarkTests: XCTestCase {
 			OCBookmarkManager.shared.addBookmark(bookmark)
 			UtilsTests.refreshServerList()
 
+			//Actions
 			EarlGrey.select(elementWithMatcher: grey_accessibilityID("server-bookmark-cell")).perform(grey_swipeFastInDirection(.left))
 			EarlGrey.select(elementWithMatcher: grey_text("Edit".localized)).perform(grey_tap())
 			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-name-name")).perform(grey_replaceText(expectedServerName))
@@ -69,6 +101,58 @@ class EditBookmarkTests: XCTestCase {
 			EarlGrey.select(elementWithMatcher: grey_text(expectedServerName)).assert(grey_sufficientlyVisible())
 
 			//Reset status
+			UtilsTests.deleteAllBookmarks()
+			UtilsTests.refreshServerList()
+		}
+	}
+
+	/*
+	 * PASSED if: Credential fields not displayed after clicking in "Delete Authentication Data"
+	 */
+	func testCheckEditDeleteAuthenticationDataBasicAuth () {
+
+		if let bookmark: OCBookmark = UtilsTests.getBookmark() {
+
+			OCBookmarkManager.shared.addBookmark(bookmark)
+			UtilsTests.refreshServerList()
+
+			//Actions
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("server-bookmark-cell")).perform(grey_swipeFastInDirection(.left))
+			EarlGrey.select(elementWithMatcher: grey_text("Edit".localized)).perform(grey_tap())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-auth-data-delete")).perform(grey_tap())
+
+			//Assert
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-username")).assert(grey_notVisible())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-password")).assert(grey_notVisible())
+
+			//Reset status
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("cancel")).perform(grey_tap())
+			UtilsTests.deleteAllBookmarks()
+			UtilsTests.refreshServerList()
+		}
+	}
+
+	/*
+	 * PASSED if: Credential fields not displayed after clicking in "Delete Authentication Data"
+	 */
+	func testCheckEditDeleteAuthenticationDataOAuth2 () {
+
+		if let bookmark: OCBookmark = UtilsTests.getBookmark(authenticationMethod: OCAuthenticationMethodIdentifier.oAuth2) {
+
+			OCBookmarkManager.shared.addBookmark(bookmark)
+			UtilsTests.refreshServerList()
+
+			//Actions
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("server-bookmark-cell")).perform(grey_swipeFastInDirection(.left))
+			EarlGrey.select(elementWithMatcher: grey_text("Edit".localized)).perform(grey_tap())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-auth-data-delete")).perform(grey_tap())
+
+			//Assert
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-username")).assert(grey_notVisible())
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("row-credentials-password")).assert(grey_notVisible())
+
+			//Reset status
+			EarlGrey.select(elementWithMatcher: grey_accessibilityID("cancel")).perform(grey_tap())
 			UtilsTests.deleteAllBookmarks()
 			UtilsTests.refreshServerList()
 		}
